requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pytesseract==0.3.13
pdf2image==1.17.0
spacy==3.7.5
rapidfuzz==3.9.6
python-dotenv==1.0.1
google-generativeai==0.7.2 # optional; comment out if not using Gemini